# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "lagi belajar asik golang sampe mampus :v",
        "title": "GO Fiber Training",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Jefri Herdi Triyanto",
            "email": "jefriherditriyanto@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
        "/api/main/v1/add": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "menambahkan product baru kedalam database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "managements"
                ],
                "parameters": [
                    {
                        "description": "Content New Product",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateProduct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success add product",
                        "schema": {
                            "$ref": "#/definitions/response.ProductAdd200"
                        }
                    },
                    "400": {
                        "description": "code harus di isi !!, body is still missing",
                        "schema": {
                            "$ref": "#/definitions/response.ProductAdd400"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.CreateProduct": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "response.ProductAdd200": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ProductAdd400": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}